Update 23 files
- /se-qubip/src/aes/aes_hw_temp.c
- /se-qubip/bit/.old/3.0/SE_QUBIP_PAU.bit
- /se-qubip/bit/.old/3.0/SE_QUBIP_PAU_2.bit
- /se-qubip/bit/.old/3.0/SE_QUBIP_PYNQZ2.bit
- /se-qubip/bit/.old/3.0/SE_QUBIP_ZCU104.bit
- /se-qubip/bit/.old/4.0/PYNQZ2_SE_QUBIP_AXI_4.0.bit
- /se-qubip/bit/.old/4.0/PYNQZ2_SE_QUBIP_EDDSA_AXI_4.0.bit
- /se-qubip/bit/.old/4.0/SE_QUBIP_PYNQZ2.bit
- /se-qubip/bit/.old/4.0/SE_QUBIP_ZCU104.bit
- /se-qubip/bit/.old/5.0/GENESYSII_SE_QUBIP_I2C_5.0.bit
- /se-qubip/bit/.old/5.0/PYNQZ2_SE_QUBIP_AXI_5.0.bit
- /se-qubip/bit/PYNQZ2_SE_QUBIP_0.9.bit
- /se-qubip/bit/GENESYSII_SE_QUBIP_0.9.bit
- /se-qubip/bit/ZCU104_SE_QUBIP_0.9.bit
- /se-qubip/rtl/aes/aes_core.v
- /se-qubip/src/aes/aes_hw.c
- /se-qubip/src/sha3/sha3_shake_hw.c
- /demo/config.conf
- /demo/Makefile
- /demo/src/demo_aes_speed.c
- /se-qubip.h
- /SE_QUBIP_0.9.rar
- /README.md

+++ se-qubip/src/sha3/sha3_shake_hw.c
@@ -31,7 +31,7 @@ void sha3_shake_interface_init(INTF interface, int VERSION) {
 	else					op_version = 2 << 2;
 
 	op = (unsigned long long int)ADD_SHA3 << 32 | ((op_version | 0) & 0xFFFFFFFF);; // RESET OFF
-	write_INTF(interface, &op, CONTROL, sizeof(unsigned int));
+	write_INTF(interface, &op, CONTROL, sizeof(unsigned long long int));
 
 }
